+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 # Changelog
+## 0.5.0 - Not released
+
+### Features
+
+### Bugfix
+
+### Miscellaneous
+- PhpOffice\PhpPowerpoint\Style\Font::setStriketrough has been removed : Use setStrikethrough
+- PhpOffice\PhpPowerpoint\AbstractShape::getSlide has been removed
+- PhpOffice\PhpPowerpoint\AbstractShape::setSlide has been removed
+- PhpOffice\PhpPowerpoint\DocumentLayout::getLayoutXmilli has been removed : getCX(DocumentLayout::UNIT_MILLIMETER)
+- PhpOffice\PhpPowerpoint\DocumentLayout::getLayoutYmilli has been removed : getCY(DocumentLayout::UNIT_MILLIMETER)
+- PhpOffice\PhpPowerpoint\DocumentLayout::setLayoutXmilli has been removed : setCX(DocumentLayout::UNIT_MILLIMETER)
+- PhpOffice\PhpPowerpoint\DocumentLayout::setLayoutYmilli has been removed : setCY(DocumentLayout::UNIT_MILLIMETER)
+
 ## 0.4.0 - 2015-07-07
 
 ### Features